+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add --allow-inexisting flag to allow inexisting buckets
+
 ### Changed
 - Upgrade PieMaker to 1.4.0